+++ mobile/android/android-components/components/service/sync-logins/src/main/java/mozilla/components/service/sync/logins/AsyncLoginsStorage.kt
@@ -15,8 +15,8 @@ import mozilla.appservices.logins.LoginsStorage
 import mozilla.appservices.logins.MemoryLoginsStorage
 import mozilla.components.concept.sync.SyncAuthInfo
 import mozilla.components.concept.sync.SyncStatus
-import mozilla.components.concept.sync.SyncableStore
 import mozilla.appservices.sync15.SyncTelemetryPing
+import mozilla.components.concept.sync.LockableStore
 
 /**
  * This type contains the set of information required to successfully
@@ -388,7 +388,7 @@ open class AsyncLoginsStorageAdapter<T : LoginsStorage>(private val wrapped: T) 
 data class SyncableLoginsStore(
     val store: AsyncLoginsStorage,
     val key: () -> Deferred<String>
-) : SyncableStore {
+) : LockableStore {
     override suspend fun sync(authInfo: SyncAuthInfo): SyncStatus {
         return try {
             withUnlocked {
@@ -397,6 +397,15 @@ data class SyncableLoginsStore(
             }
         } catch (e: LoginsStorageException) {
             SyncStatus.Error(e)
+        }
+    }
+
+    override suspend fun <T> unlocked(encryptionKey: String, block: (store: LockableStore) -> T): T {
+        return try {
+            store.ensureUnlocked(encryptionKey).await()
+            block(this)
+        } finally {
+            store.ensureLocked().await()
         }
     }